Volunteer page

See merge request com3420-2020-21/team21/project!15

+++ db/migrate/20210403204845_create_prospective_donors.rb
@@ -3,7 +3,7 @@ class CreateProspectiveDonors < ActiveRecord::Migration[6.0]
     create_table :prospective_donors do |t|
       t.string :name
       t.string :type
-      t.integer :contactnumber
+      t.bigint :contactnumber
       t.string :internalcontactlink
       t.string :email
       t.timestamps
